Module 3 changes to be moved to relevant branch
