+++ .env.test
@@ -1,0 +1,13 @@
+# Flask server configuration.
+FLASK_APP=app
+FLASK_ENV=development
+FLASK_SKIP_DOTENV=True
+
+# Change the following values for local development.
+SECRET_KEY=secret-key
+
+# Fake Trello API Key And Token
+API_KEY=358x5xy7xyxy96z686983250v1b8ab8c7
+TOKEN=mum61b75051vc88c69c6ead781e54gg885b16d8394bzz5bcb223a48tage19507
+BOARD_ID=5f5a4b008a129438843fcf0f
+LIST_ID=5f5a4b008a129438843fcf10
